ability to replace SourceFinder [WIP]
