+++ packages/EasyCodingStandard/src/DependencyInjection/CompilerPass/CollectorCompilerPass.php
@@ -8,8 +8,7 @@ use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symplify\EasyCodingStandard\Contract\Finder\ExtraFilesProviderInterface;
-use Symplify\EasyCodingStandard\Finder\SourceFinder;
+use Symplify\EasyCodingStandard\Finder\SourceFinderInterface;
 use Symplify\EasyCodingStandard\FixerRunner\Application\FixerFileProcessor;
 use Symplify\EasyCodingStandard\SniffRunner\Application\SniffFileProcessor;
 use Symplify\PackageBuilder\Adapter\Symfony\DependencyInjection\DefinitionCollector;
@@ -21,7 +20,7 @@ final class CollectorCompilerPass implements CompilerPassInterface
         $this->collectCommandsToConsoleApplication($containerBuilder);
         $this->collectFixersToFixerFileProcessor($containerBuilder);
         $this->collectSniffsToSniffFileProcessor($containerBuilder);
-        $this->collectExtraFilesProvidersToSourceFinder($containerBuilder);
+        $this->useCustomSourceFinder($containerBuilder);
     }
 
     private function collectCommandsToConsoleApplication(ContainerBuilder $containerBuilder): void
@@ -54,14 +53,8 @@ final class CollectorCompilerPass implements CompilerPassInterface
         );
     }
 
-    private function collectExtraFilesProvidersToSourceFinder(ContainerBuilder $containerBuilder): void
+    private function useCustomSourceFinder(ContainerBuilder $containerBuilder): void
     {
-    	// TODO: add Symplify\EasyCodingStandard\Finder\DefaultFilesProvider if there are no ExtraFilesProviderInterface
-        DefinitionCollector::loadCollectorWithType(
-            $containerBuilder,
-            SourceFinder::class,
-            ExtraFilesProviderInterface::class,
-            'addExtraFilesProvider'
-        );
+    	// ???
     }
 }
