+++ packages/EasyCodingStandard/src/Finder/SourceFinder.php
@@ -2,21 +2,11 @@
 
 namespace Symplify\EasyCodingStandard\Finder;
 
+use Symfony\Component\Finder\Finder;
 use SplFileInfo;
-use Symplify\EasyCodingStandard\Contract\Finder\ExtraFilesProviderInterface;
 
-final class SourceFinder
+final class SourceFinder implement SourceFinderInterface
 {
-    /**
-     * @var ExtraFilesProviderInterface[]
-     */
-    private $extraFilesProvider = [];
-
-    public function addExtraFilesProvider(ExtraFilesProviderInterface $sourceProvider): void
-    {
-        $this->extraFilesProvider[] = $sourceProvider;
-    }
-
     /**
      * @param string[]
      * @return SplFileInfo[]
@@ -25,10 +15,39 @@ final class SourceFinder
     {
         $files = [];
 
-        foreach ($this->extraFilesProvider as $extraFilesProvider) {
-            $files = array_merge($files, $extraFilesProvider->provideForSource($source));
+        foreach ($source as $singleSource) {
+            if (is_file($singleSource)) {
+                $files = $this->processFile($files, $singleSource);
+            } else {
+                $files = $this->processDirectory($files, $singleSource);
+            }
         }
 
         return $files;
+    }
+
+    /**
+     * @param SplFileInfo[] $files
+     * @return SplFileInfo[]
+     */
+    private function processFile(array $files, string $file): array
+    {
+        return array_merge($files, [
+            $file => new SplFileInfo($file),
+        ]);
+    }
+
+    /**
+     * @param SplFileInfo[] $files
+     * @param string $file
+     * @return SplFileInfo[]
+     */
+    private function processDirectory(array $files, string $directory): array
+    {
+        $finder = (new Finder)->files()
+            ->name('*.php')
+            ->in($directory);
+
+        return array_merge($files, iterator_to_array($finder->getIterator()));
     }
 }
