+++ packages/EasyCodingStandard/src/Finder/SourceFinderInterface.php
@@ -1,0 +1,14 @@
+<?php declare(strict_types=1);
+
+namespace Symplify\EasyCodingStandard\Finder;
+
+use SplFileInfo;
+
+interface SourceFinderInterface
+{
+    /**
+     * @param string[]
+     * @return SplFileInfo[]
+     */
+    public function find(array $source): array;
+}
